Fixing processing and gui bugs-12/17

+++ src/main/java/MicroProtocol/listeners/ChangeThresholdListener.java
@@ -23,6 +23,6 @@ package MicroProtocol.listeners;
  */
 public interface ChangeThresholdListener {
     
-    public void thresholdChanged(int min, int max);
+    public void thresholdChanged(double min, double max);
     
 }
